+++ src/inscription-parser.service.helper.ts
@@ -1,4 +1,5 @@
 import { MAX_DECOMPRESSED_SIZE_MESSAGE, brotliDecode } from "./brotli-decode";
+import { byteArrayToHex } from "./lib/conversions";
 
 /**
  * Bitcoin Script Opcodes
@@ -186,44 +187,6 @@ export function hasInscription(witness: string[]): boolean {
 }
 
 /**
- * Converts a byte array to a hexadecimal string.
- *
- * @param byteArray - The array of bytes to convert.
- * @returns The hexadecimal string representation of the byte array.
- */
-export function byteArrayToHex(byteArray: Uint8Array): string {
-  return Array.from(byteArray)
-              .map(byte => byte.toString(16).padStart(2, '0'))
-              .join('');
-}
-
-/**
- * Extracts the parent inscription ID from a field in an inscription.
- * The parent field value consists of a 32-byte transaction ID (TXID) followed by a four-byte little-endian index.
- * The TXID part is reversed in order, and trailing zeroes are omitted.
- *
- * @param parentField - The field containing the parent inscription data.
- * @returns The parent inscription ID as a string, or undefined if the parent field is not provided.
- */
-export function extractParent(value: Uint8Array | undefined): string | undefined {
-
-  if (value === undefined) {
-    return undefined;
-  }
-
-  // Reverse the TXID part and convert it to hexadecimal
-  const txId = value.slice(0, 32).reverse();
-  const txIdHex = byteArrayToHex(txId);
-
-  // Convert the 4-byte little-endian index to a decimal number
-  const indexBytes = value.slice(32, 36); // Get the index part
-  const index = littleEndianBytesToNumber(indexBytes);
-
-  // Combine TXID and index to form the parent inscription ID
-  return txIdHex + 'i' + index;
-}
-
-/**
  * Extracts the pointer value from a given field in an inscription.
  * The pointer value is a little-endian encoded integer specifying the sat position in the outputs.
  *
@@ -276,4 +239,29 @@ export function brotliDecodeUint8Array(bytes: Uint8Array): Uint8Array {
     }
     throw error;
   }
+}/**
+ * Extracts the parent inscription ID from a field in an inscription.
+ * The parent field value consists of a 32-byte transaction ID (TXID) followed by a four-byte little-endian index.
+ * The TXID part is reversed in order, and trailing zeroes are omitted.
+ *
+ * @param parentField - The field containing the parent inscription data.
+ * @returns The parent inscription ID as a string, or undefined if the parent field is not provided.
+ */
+export function extractParent(value: Uint8Array | undefined): string | undefined {
+
+  if (value === undefined) {
+    return undefined;
+  }
+
+  // Reverse the TXID part and convert it to hexadecimal
+  const txId = value.slice(0, 32).reverse();
+  const txIdHex = byteArrayToHex(txId);
+
+  // Convert the 4-byte little-endian index to a decimal number
+  const indexBytes = value.slice(32, 36); // Get the index part
+  const index = littleEndianBytesToNumber(indexBytes);
+
+  // Combine TXID and index to form the parent inscription ID
+  return txIdHex + 'i' + index;
 }
+
