+++ src/inscription-parser.service.ts
@@ -1,4 +1,5 @@
-import { OP_0, OP_ENDIF, brotliDecodeUint8Array, extractParent, extractPointer, getKnownFieldValue, getNextInscriptionMark, knownFields, readPushdata } from "./inscription-parser.service.helper";
+import { OP_0, OP_ENDIF, brotliDecodeUint8Array, extractPointer, getKnownFieldValue, getNextInscriptionMark, knownFields, readPushdata } from "./inscription-parser.service.helper";
+import { extractParent } from "./inscription-parser.service.helper";
 import { binaryStringToBase64, hexStringToUint8Array, uint8ArrayToSingleByteChars, utf8BytesToUtf16String } from "./lib/conversions";
 import { ParsedInscription } from "./parsed-inscription";
 import { CBOR } from "./cbor";
